NLP/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
scikit-learn==1.3.2
nltk==3.8.1
textstat==0.7.3
numpy==1.26.2
pytest==7.4.3
